refactor: modularize common functionality
